% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mc_fit}
\alias{mc_fit}
\title{Fit a mixture cure model}
\usage{
mc_fit(spec, control = list())
}
\arguments{
\item{spec}{A mc_spec object}

\item{control}{Engine-specific control list}
}
\description{
Fit a mixture cure model
}
