% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_specs.R
\name{mc_spec}
\alias{mc_spec}
\title{Create a mixture cure model specification}
\usage{
mc_spec(
  incidence,
  latency,
  data,
  time = "time",
  status = "status",
  engine = c("mixcuref", "ClipMixcure", "mixcuref.cr"),
  ...
)
}
\arguments{
\item{incidence}{Formula for cure probability}

\item{latency}{Formula for susceptible survival}

\item{data}{Data frame}

\item{time}{Name of time variable}

\item{status}{Name of event indicator (1=event, 0=censored)}

\item{engine}{One of "mixcuref", "ClipMixcure", "mixcure.cr"}
}
\description{
Create a mixture cure model specification
}
\examples{
library(Mixcuref.Meta)
# for MC model fitting with mixcuref() function
library(mixcuref)
data(ANNbcBMdat1)
spec <- mc_spec(
  incidence = ~ Her2,     # cure probability model
  latency   = ~ Her2,     # survival model
  data      = ANNbcBMdat1,
  time      = "Time",
  status    = "CENS",
  engine    = "mixcuref"
)

## mc_fit() sees engine = "mixcuref"
## Meta-layer adapter builds Surv(time, status) ~ age + stage + trt
## And calls:
## mixcuref::mixcure.penal.est(...)
## Returns the fitted mixcure object, wrapped in a meta-layer container

fit <- mc_fit(
  spec,
  control = list(
    pl     = TRUE,     # penalised likelihood
    maxit  = 200
  )
)
fit
summary(fit)

library(ClipMixcure)
# wrap to clip.mixcure()
data(ANNbcBMdat5_miss)
library(mice)

spec_clip <- mc_spec(
  incidence = ~ age + trt + x2,
  latency   = ~ age + trt + x1 + x2,
  data      = dat2,
  time      = "time",
  status    = "status",
  engine    = "ClipMixcure"
)

fit2  <- mc_fit(spec2,
control = list(
    method = "clip",
    maxit  = 200
  )
)

fit2
summary(fit2)

library(mixcuref.cr)
spec_plcr <- mc_spec(
  incidence = ~ Her2,     # cure probability model
  latency   = ~ Her2,     # survival model
  data      = ANNbcBMdat1,
  time      = "Time",
  status    = "CENS",
  engine    = "mixcuref.cr"
)

fit_plcr <- mc_fit(
  spec_plcr,
  control = list(
    pl   = FALSE,  # example flag (depends on your engine)
    apct = 0.05,   # example argument used in your PLCR code
    maxit = 200
  )
)

fit_plcr
summary(fit_plcr)
coef(fit_plcr)


}
